# api/api.yaml
openapi: 3.0.0
info:
  description: API for the TeamPickr app
  version: '1.0.0'
  title: TeamPickr
  contact:
    email: david.saltares@gmail.com
  license:
    name: MIT
    url: 'https://github.com/dsaltares/teamfindr/blob/main/LICENSE'
paths:
  /auth/verify:
    get:
      summary: verifies that the session for the user is valid
      operationId: verifyUser
      responses:
        '200':
          description: authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifySuccessResponse'
        '401':
          description: the session for the user is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/push:
    post:
      summary: registers for push notifications
      operationId: postPush
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPushRequest'
      responses:
        '201':
          description: registered device for push notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPushResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    VerifySuccessResponse:
      type: object
      required:
        - user
        - pushPublicKey
      properties:
        user:
          $ref: '#/components/schemas/User'
        pushPublicKey:
          type: string
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PostPushRequest:
      type: object
      required:
        - subscription
      properties:
        subscription:
          $ref: '#/components/schemas/PushSubscription'
    PostPushResponse:
      type: object
...
    User:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        avatar:
          type: string
...
    PushSubscription:
      type: object
      required:
        - endpoint
        - keys
      properties:
        endpoint:
          type: string
        expirationTime:
          type: number
          nullable: true
        keys:
          type: object
          required:
            - p256dh
            - auth
          properties:
            p256dh:
              type: string
            auth:
              type: string